Fix: 체크박스 정상 동작하도록 수정
체크박스 단일 선택 되도록 구현,
체크된 박스 해제 가능하도록 수정
그 외, CSS 수정
