+++ src/components/Nav/NavBottomContainer/Dropdown/DropdownContent/ColorPallette/ColorPallette.js
@@ -7,7 +7,11 @@ const ColorPallette = () => {
     <div className="colorPalette">
       <div className="colorBoxContainer">
         {COLOUR_MOCK_DATA.map(box => (
-          <Link key={box.id} to={`/category/2/type/${box.id}`}>
+          <Link
+            key={box.id}
+            to={`/category/2/type/${box.id}`}
+            className="linkComponent"
+          >
             <ColorBox
               key={box.id}
               color={box.color}
@@ -19,7 +23,11 @@ const ColorPallette = () => {
       </div>
       <div className="colorBoxContainer">
         {COLOUR_MOCK_DATA2.map(box => (
-          <Link key={box.id} to={`/category/2/type/${box.id}`}>
+          <Link
+            key={box.id}
+            to={`/category/2/type/${box.id}`}
+            className="linkComponent"
+          >
             <ColorBox
               key={box.id}
               color={box.color}
