+++ src/pages/Main/Main.js
@@ -28,6 +28,7 @@ function Main() {
         throw new Error(`에러가 발생했습니다. ${error.message}`);
       }
     };
+
     const options = {
       method: 'GET',
       headers: {
