+++ src/pages/ProductsList/FilterMenu/FilterMenu.js
@@ -1,11 +1,30 @@
 import { useState } from 'react';
-import Filter from './Filter/Filter';
 import './FilterMenu.scss';
 
 function FilterMenu({ title, list, handleChangeFilter }) {
   const [isMenuOpened, setIsMenuOpened] = useState(false);
+  const [checkedBox, setCheckedBox] = useState('');
 
   const handleClickOpenBtn = e => setIsMenuOpened(prev => !prev);
+
+  const handleCheckBox = e => {
+    clickCheck(e.target);
+    handleChangeFilter(e);
+  };
+
+  function clickCheck(target) {
+    document
+      .querySelectorAll(`input[type=checkbox]`)
+      .forEach(el => (el.checked = false));
+
+    if (checkedBox === target.value) {
+      setCheckedBox('');
+      return;
+    }
+
+    target.checked = true;
+    setCheckedBox(target.value);
+  }
 
   return (
     <div key={title} className={`filterMenu ${isMenuOpened ? 'openMenu' : ''}`}>
@@ -19,13 +38,22 @@ function FilterMenu({ title, list, handleChangeFilter }) {
       </div>
       {list.map(subMenu => {
         return (
-          <Filter
+          <div
+            className={`filter ${isMenuOpened ? 'active' : ''}`}
             key={subMenu}
-            handleChangeFilter={handleChangeFilter}
-            isMenuOpened={isMenuOpened}
-            subMenu={subMenu}
-            title={title}
-          />
+          >
+            <input
+              type="checkbox"
+              className="check"
+              placeholder={subMenu}
+              value={subMenu}
+              name="orderCheckBox"
+              onChange={e => handleCheckBox(e)}
+              data-filtertype={title}
+            />
+            <label htmlFor="check" />
+            <span>{subMenu}</span>
+          </div>
         );
       })}
     </div>
