+++ src/pages/ProductsList/ProductsList.scss
@@ -36,6 +36,10 @@
           border: 1px solid black;
           border-radius: 1rem;
           cursor: pointer;
+
+          i {
+            margin-left: 0.5rem;
+          }
         }
       }
